// app/models/matchesFromPlayerResponse.tsx
export interface MatchesResponse {
  status: number;
  data: Datum[];
}

export interface Datum {
  is_available?: boolean;
  metadata: Metadata;
  players: Players;
  observers: any[];
  coaches: Coach[];
  teams: Teams;
  rounds: Round[];
  kills: Kill[];
}

export interface Coach {
  puuid: string;
  team: Team;
}

export enum Team {
  Blue = "Blue",
  Red = "Red",
}

export interface Kill {
  kill_time_in_round: number;
  kill_time_in_match: number;
  round?: number;
  killer_puuid: string;
  killer_display_name: DisplayName;
  killer_team: Team;
  victim_puuid: string;
  victim_display_name: DisplayName;
  victim_team: Team;
  victim_death_location: Location;
  damage_weapon_id: DamageWeaponID;
  damage_weapon_name: DamageWeaponNameEnum | null;
  damage_weapon_assets: DamageWeaponAssetsClass;
  secondary_fire_mode: boolean;
  player_locations_on_kill: PlayerLocationsOn[];
  assistants: Assistant[];
}
// ...
export interface Metadata {
  map: string;
  game_version: string;
  game_length: number;
  game_start: number;
  game_start_patched: string;
  rounds_played: number;
  mode: string;
  mode_id: string;
  queue: string;
  season_id: string;
  platform: PlatformEnum;
  matchid: string;
  premier_info: PremierInfo;
  region: string;
  cluster: string;
}

export enum PlatformEnum {
  PC = "PC",
}

export interface PremierInfo {
  tournament_id: null;
  matchup_id: null;
}

export interface Players {
  all_players: AllPlayer[];
  red: AllPlayer[];
  blue: AllPlayer[];
}
// ...
export interface Round {
  winning_team: Team;
  end_type: EndType;
  bomb_planted: boolean;
  bomb_defused: boolean;
  plant_events: PlantEvents;
  defuse_events: DefuseEvents;
  player_stats: PlayerStat[];
}
// ...
export interface Teams {
  red: Blue;
  blue: Blue;
}

export interface Blue {
  has_won: boolean;
  rounds_won: number;
  rounds_lost: number;
  roster: null;
}